Bugfix: deleteDirectory won't delete targets of symlinks
In its current state, `YellowToolbox::deleteDirectory()` does not delete any symlink in the directory. Instead, it tries to delete the *target* of the symlink, because of `getRealPath()`. If the target is a file or an empty dir, then it's deleted and all that's left in the directory is a broken link (and this broken link prevents the dir to be removed by `rmdir()`).

This commit fixes that behaviour by deleting the *link*, keeping the target intact. The directory is then removed with `rmdir()` as expected.